Use printf instead of echo for consistency
Turns out POSIX just left backslash as implementation defined for echo.

Many versions of echo have -e/-E switches to decide if escape sequences
are replaced or not, but not all. In the case of dash, which is /bin/sh
on some Debian systems, escape sequences can't be turned off.

printf, on the other hand, does have specified behaviour in that it
will expand escape sequences in the format string, but not the
%s arguments.