Automate headless compilation of Forth source.
base.fs adds an instruction to jam the CPU,
which is connected to exiting in Vice.
x64sc has more accurate video simulation, which we don't need.
A timeout in cycles is added so a build can fail,
and a final screenshot so build result can be checked.
That screenshot may or may not show the final "ok!"
because the exit can happen faster than the display update.